PlenOctree/requirements-colab.txt
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
imageio==2.31.6
imageio-ffmpeg==0.4.9
ipdb==0.13.13
lpips==0.1.4
jax==0.4.20
jaxlib @ https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.20+cuda12.cudnn89-cp310-cp310-manylinux2014_x86_64.whl#sha256=4cbab7cf74f67e34dc3b71126a16e8286df79ab3a98ed8fd0bdd9fe1de2724aa
flax==0.7.5
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
Pillow==9.4.0
PyYAML==6.0.1
tensorflow==2.8.1
PyMCubes==0.1.4
svox==0.2.32
tqdm==4.66.1
gpustat==1.1.1
optax==0.1.7
orbax-checkpoint==0.4.4
pandas==1.5.3
numpy==1.23.5

--extra-index-url https://download.pytorch.org/whl/cu121
pytorch==2.1.0
torchvision==0.16.0 
torchaudio==2.1.0 